Initial theory
Some of my ideas about learning before starting the course, as well as ways I built on these in the first set of readings, assigned for the second week of the course.

In other words, this is my "week 2 reflection", seeded with some initial ideas (which we all discussed with several different partners at the beginning of class).

It's not particularly organized or cohesive yet but it's a start :)

(The class I am referring to is Educational Psychology at Swarthmore College taught by Ann Renningeri)